LPS-196947 Inline PanelAppMyAccountPermissions methods and fields into com.liferay.application.list.my.account.permissions.internal.model.listener.CompanyModelListener and inline PanelAppServiceTrackerCustomizer in com.liferay.application.list.user.personal.site.permissions.internal.model.listener.CompanyModelListener
